Adding support for --credentials
To enable the use of an external credentials provider, without
the need to integrate with specific APIs, the `--credentials`
flags is being added. The end-user is then able to provide
the source file however they wish, so long as it adheres to the
following structure:

```json
{
    "username": string,
    "password": string,
    "agentpassword": string
}
```

Currently, only readable files are supported, but support for
execution may be added at a later date.

* Added `commands.Credentials` along with its receiver functions
  `Marshal`, `Unmarshal` and `ReadFromSource`
* Refactored `commands.ReadFile` to use `content` instead of
  `certificate` to avoid confusion
* Added `commands.TestCredentials`
* Added `Credentials` to `commands/management/addMySQLCommand` to
  store the path to the credentials file
* Added `commands/management/addMySQLCommand.GetCredentials`
  receiver function to enable process the specified source. This is
  called whenever `--credentials` is set to a non-zero-length string
  and this can then be used in place of `--username`, `--password`,
  and `--agent-password`